# requirements/reference/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StepEdit": {
        "properties": {
          "is_checkpoint": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Checkpoint"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "skill_node_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Node Id"
          }
        },
        "title": "StepEdit",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "mathbank-rest",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/attempts/{attempt_id}/diagnoses": {
      "get": {
        "description": "Internal (X-Admin-Api-Key): full diagnoses with failure modes, confidences and evidence.",
        "operationId": "admin_list_diagnoses_v1_admin_attempts__attempt_id__diagnoses_get",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Admin List Diagnoses V1 Admin Attempts  Attempt Id  Diagnoses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin List Diagnoses",
        "tags": [
          "step-runtime"
        ]
      }
    },
...
    "/v1/admin/imports/steps/{step_id}": {
      "patch": {
        "description": "Adjust skill / mark checkpoint. Emits SOLUTION_STEP_CHANGED (graph + embedding re-projection).",
        "operationId": "edit_step_v1_admin_imports_steps__step_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "step_id",
            "required": true,
            "schema": {
              "title": "Step Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepEdit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Edit Step V1 Admin Imports Steps  Step Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Step",
        "tags": [
          "admin-imports"
        ]
      }
    },
    "/v1/admin/knowledge-gaps": {
      "get": {
        "description": "Internal (X-Admin-Api-Key): knowledge gaps across students \u2014 status totals, the most common open\ntargets, and recent hypotheses with their diagnosis and latest recovery plan.",
        "operationId": "admin_gap_overview_v1_admin_knowledge_gaps_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "UNRESOLVED",
                    "CONFIRMED",
                    "REJECTED",
                    "RESOLVED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "description": "e-mail substring or student uuid",
            "in": "query",
            "name": "student",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "e-mail substring or student uuid",
              "title": "Student"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Gap Overview V1 Admin Knowledge Gaps Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Gap Overview",
        "tags": [
          "step-runtime"
        ]
      }
    },
...
    "/v1/admin/pipeline/runs": {
      "get": {
        "operationId": "pipeline_runs_v1_admin_pipeline_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Pipeline Runs V1 Admin Pipeline Runs Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pipeline Runs",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/recovery-plans": {
      "get": {
        "description": "Internal (X-Admin-Api-Key): recovery plans with item outcome counts.",
        "operationId": "admin_recovery_plans_v1_admin_recovery_plans_get",
        "parameters": [
          {
            "in": "query",
            "name": "student_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ACTIVE",
                    "SUSPENDED",
                    "COMPLETED",
                    "EXHAUSTED",
                    "ABORTED",
                    "SUPERSEDED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Admin Recovery Plans V1 Admin Recovery Plans Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Recovery Plans",
        "tags": [
          "step-runtime"
        ]
      }
    },
...
    "/v1/problem-images/{image_id}": {
      "get": {
        "operationId": "get_problem_image_v1_problem_images__image_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Problem Image",
        "tags": [
          "step-runtime"
        ]
      }
    },
...
  }
}